requirements.txt
pyTelegramBotAPI==4.16.1
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
deep-translator==1.11.4
